% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_tools.R
\name{plotArrayAsImage}
\alias{plotArrayAsImage}
\title{Plot a 3D array as an RGB image}
\usage{
plotArrayAsImage(rgb.array, main = "")
}
\arguments{
\item{rgb.array}{3D RGB array with R, G, and B channels (pixel rows x pixel
columns x color channels) to be plotted in the plot window as an actual
image.}

\item{main}{Optional title to display for image.}
}
\description{
Plots a 3D array as an RGB image in the plot window.
}
\examples{
# Read in image
flowers <- jpeg::readJPEG(system.file("extdata", "flowers.jpg", package =
"countcolors"))

# Plot
plotArrayAsImage(flowers, main = "flowers!")

}
